Ask-on-Entry: eliminated over-use of confirmation dialog.
We only want this if people get as far as submitting the results without
asking the questions, so it should be part of the submission flow, not
part of the Ask-on-Entry question flow. Removed the unnecessarily generic
parts of the code and moved it to be an internal component of the page
where it is actually used.