Refactor GraphQL into autowired resolvers
Splitting into the resolver types makes it easy to add the required security annotations to the various query endpoints.

Removed the GraphqlConfiguration class as it's no longer needed. We're now getting type-checked query support at application startup, so we'll get early warning if we introduce a type or query that isn't supported.