Disentangled conflicting bean profiles.
Made the security-disabling feature into its own profile, and included
it in the "dev" and "cloud" profiles so that it is used in local dev
and in cloud.gov deploys, but other local dev settings are not.

Also added fallback placeholders for the Okta OAuth2 client ID and
secret, so that failure to supply them does not crash the app.