Added "whoami" resolver, with service calls back to DB.
Massaged the Person model slightly to allow the use case where we have
just the basic name information, and made the API User model wrap the
ApiUser model from the database (oof) to allow fetching the required
information about the current user and organization in a single call.

Also learned a little about entity-graph management in JPA, the hard way.