Added tests for patient-list fetch.
Somewhat laboriously added two different test cases, one of which should have failed
but initially did not, and the other of which kindly did fail continuously, for
the problem caused by mis-defining the "street" column in an address.

Also imported the database-truncating tool to wipe the DB during tests, making
it even more imperative to eventually stop using the dev DB for testing.