Make React routes relative to public URL
Addresses an issue where the frontend was not loading correctly in Azure, due to the basepath being set to / rather than the relative context it requires.

Simple fix to add the PUBLIC_URL env variable to the router basename.

Also updated the deploy commands to insert the correct parameters when deploying to dev and prod.