Pulled TestResult and DeviceType up into BaseTestInfo.
Since TestResult is nullable in one case but not the other, used some
JPA sorcery on TestEvent (and NoJsonTestEvent) to enforce that.

Since there are outstanding issues (CDCgov/prime-central#151 and #75)
that hinge on the fact that we are not saving the device type when we
create a TestOrder, made DeviceType mandatory for a TestOrder, and
updated a bunch of tests and construction code to make that fly.

Also updated the annotations on a couple of immutable join columns
to make sure that hibernate understands that they are immutable.