Fix 404 on Okta logout
Addressed 2 issues in the code.

1. The Okta logout URL was incorrect, we needed to point to the default authorization server.
My fault on sending along bad info.

2. We need to pass in the id_token along with the request.
Modified the index file to capture the id token alongside the access token and store both in local storage.

On logout, we fetch the id_token from localstorage and send it along.
We also remove the tokens from localStorage before doing the Okta redirect.