Fixed github action cache issues from node upgrade.
Made the node cache key include the node version, and made the frontend
jobs all matrixed so that they can be run in 12 and 14 simultaneously
now, and 16 in the future, without conflicts. Since deploy isn't a
thing that makes sense as a matrix, just made the node version an
environment variable and made the cache key include it.

Also made the operating system matrixed, just to facilitate testing
the new ubuntu version: we will probably switch it back to a
single-element dimension of "ubuntu-latest" in the future.

Also deleted the unused deployTraining workflow, because it's unused.