Gave service tests their own slice configuration.
Configured testing of services with everything below the web
layer wired up, instead of doing a weird frankenstein based on
the JPA slice test configuration.

This required (at least, doing it this way required) explicitly
disabling the web security adapter, but it wasn't doing anything
useful in this context anyway.