Updated group claims and group names for IDP-based authorization.
Badged all groups with a leading "SR-" to indicate that they belong
to SimpleReport, then differentiated from there. Since the production,
staging and dev/test admin groups can actually be distinct and can fit
into this model, made the group claims match only on the environment
portion (e.g. "SR-PROD-"), with the "-TENANT:" matching to be done
in Java code.