Handle a weird edge case in user configuration.
If a user is allowed to log in but has no valid organization role and is not
a super-admin, the whoami endpoint should not fail (though, admittedly, this
just kicks the can down the road to the app).

Also incidentally refactored User construction to placate Sonar, and tidied
up some API tests.