Updated specimen-type migration to run in all environments.
Since there is no uniqueness constraint on facility_device_type, it has acquired
duplicate rows in production, which caused the migration to fail when we copied
the rows of that table one-for-one into a new table that does have a uniqueness
constraint. Changed the migration to filter out duplicates, and updated the
changelog to allow both migrations to be considered valid.