Added fixup migration to fix upper environments.
Some small number of test_order rows have nulls in a column we
thought was filled with actual values. Based on the real-world
shape of this data, it is known to be safe to populate these rows
with the default value they should have picked up from the
facility or organization, had we been running a later version
of the application when they were created.