Updated specimen-type migration to run in all environments. (#905)
* Updated specimen-type migration to run in all environments.

Since there is no uniqueness constraint on facility_device_type, it has acquired
duplicate rows in production, which caused the migration to fail when we copied
the rows of that table one-for-one into a new table that does have a uniqueness
constraint. Changed the migration to filter out duplicates, and updated the
changelog to allow both migrations to be considered valid.

* Added fixup migration to fix upper environments.

Some small number of test_order rows have nulls in a column we
thought was filled with actual values. Based on the real-world
shape of this data, it is known to be safe to populate these rows
with the default value they should have picked up from the
facility or organization, had we been running a later version
of the application when they were created.

* Added separate linux ValidCheckSum entry for changed migration.

Things I had happily let slip from my mind: some (but not all) liquibase
migration checksums are platform-dependent.