Refactor Patient Experience to use REST (#813)
In order to separate the security configuration for anonymous Patient Experience
users and our authenticated SimpleReport users, this commit adds a new
PatientExperienceController and relevant SecurityConfiguration. It is protected 
with a @PreAuthorize annotation which requires that the anonymous user provide 
a valid PatientLink id and the DOB of the corresponding Patient.

The Person object for the related patient is then added to the request context
via the CurrentPatientContextHolder.

This refactor enables CORS for local development in DevSecurityConfiguration,
and disables CSRF for all environments. We are protected from exposure to CSRF 
because we don't use sessions in the first place.

A number of UI improvements are included here, which are further described
and linked in PR #813.