Support patients in getCurrentUser (#979)
Prior to this commit, the Patient Experience throws 500s in Okta-enabled 
environments because AuditingConfiguration relies on ApiUserService to
supply the current user. Patients, having not authenticated through Okta,
could not be identified and threw null pointer exceptions up the stack.

This commit adds support in SecurityConfiguration's getRealIdentity and
ApiUserService's getCurrentApiUser to support patients who have already
validated their PatientLink with a correct date of birth.