Made tests use more realistic authorization.
Made BaseApiTest pass the user in for DemoUserAuthorizationService to find in
the request, rather than just mocking all the authorization and authentication
code, and created a new set of users with no default user. Removed the special
configuration for PatientExperienceControllerTest, since it no longer does
anything special, and fixed a bunch of problems with tests having initialization
code that needed the default user in order to run.