Made demo environment authorization updateable through the API.
Replaced static configuration-driven authorization repository with one
that is driven by DemoOktaRepository, which is initialized with the same
data at startup but can be modified by the API for better testing and
better demos.

Made no-okta-mgmt profile part of the local-development profile, because
attempting to run fake security by calling the real Okta API was just
not a thing I wanted to contemplate.