Modest instrumentation refactor.
Made QueryLoggingInstrumentation a @Component, and did away with the
configuration class whose only job was to create it and add it to a
list, both of which are things the Spring ApplicationContext will
happily do for us.

Also neurotically saved the stringification of a UUID instead of letting
it happen twice in the same method.