New implementation of REST auditing.
We need to attempt the audit logging before the response is committed,
which means it cannot be done in the postHandle method. Moving to a
@PostAuthorize annotation with a custom service fixes that issue, and
issuing a custom exception so that we do not get into an exception-throwing
loop handles the issue of trying to store an audit record for an audit-logging
failure.