Big model refactor around structured names and addresses.
Added a structured "name" object to each API model that has a person's
name in it already as first/last/middle, using a certain amount of
refactoring of API models and service models, and a certain amount of
interface tagging.

Similarly, restructured the way that API models find the individual fields
of addresses, without changing the presented graphql type.