Added structured person-name input to relevant mutations.
Added a NameInput that looks a lot like NameInfo (which looks a lot like
PersonName, to be fair). Used it in all the User mutations that take a
first/middle/last name, and pushed it through to the service layer with
a (reasonably exhaustively tested) translator utility that handles both
old and new argument styles.