Add LastTestResults DataLoader
In order to get this first pass working, I've replaced `date_added`,
which was computed as test_event->test_order.created_at with simply
test_event.created_at. These are not entirely equivalent, but part
of the N+1 issue at hand was this deeply nested query

Will try to figure out why we needed the timestamp from the ORDER
instead of the EVENT... and we don't? I think we're good