Updated app_gateway configuration to include path mapping.
Added rewrite rules back into terraform configuration for the application
gateway, and told terraform to stop ignoring changes to the many related
attributes of the application gateway. Everything now seems to be set up
in a way that the documentation suggests would produce the correct result,
which makes it annoying that it does not.