Eliminated query-string passthrough for app, and commented.
Since there isn't actually a good use case for passing the query string through
to the backing blobstore on static file requests, we can eliminate it explicitly
(which would otherwise be done implicitly). Added comments to all three rewrite
rules that have URL blocks, for the benefit of future generations.