add data loaders for self registration links
This addresses several n+1 issues when getting organization list.
The query could, at worst, be 4n+2 which was causing problems when
loading the list of organizations that was used in the admin panel.

This change doesn't fully honor the graphql schema, since it does
not allow the facilities within an organization to be returned, but
it is currently not used by the frontend.

A data loader for facilities in an organization is desirable, but
this is non-trivial because facilities are used both for the
list of facilities in an organization and for the list of facilities
accessible for a user within an organization.  This compicates the
data loader since the facility list can be different depending on
context.